Added a debug framework in order to compile easily multiple point concepts.
This declarations will be used only for testing purposes and in order to solve the
compilation error. Once the compilation error is solved we should clean the code and go back to
use gtsam::Pose2.

If we try to map the gtsam_polygon_set
concept to a booost::polygon::polygon_set concept
we will get a compilation error. We don't have
a solution for this yet.

The error is :
 error: no match for ‘operator=’
 (operand types are ‘gtsam::Point2’ and  ‘const boost::polygon::point_data<double>’)
 in /usr/include/c++/7/bits/stl_algobase.h:294:16:  *__result = *__first;

This seems to indicate that the Point concept is not mapped properly.

Because is easier to debug and to go back and forward beetween gtsam::Poin2 and
point_data<double>  we will use that data types until we solve this issue.
Once we solve this we should be able to fast forward  to use gtsam::Pose2 data type.

Use : comment or or uncomment the next lines  if both are commented the default value for
the concept type will be the boost::polygon::point_data<double>