Added a new SimplePoint data type. Please read.
Now things became so wierd and difficult to read.
I know, but is the only way I've found to make
this actually work and build for the 4 Point
data types we are trying to map into boost polygon.

You shouldn't care too much about the crazy implementation
I've made for this. Basically you only need to know there
is 4 POINT_CONCEPT:
 * gtl::point_data<double> [default]
 * SimplePoint
 * gtsam::Pose2
 * gtsam::Point2

 The way to select witch point will be used to build is
 uncommenting one of the following lines(if every line
 is commented is beacuse you will use the default Point
 type, that is actually the one that boost::polygon uses)
  * #define USE_GTSAM_POINT_CONCEPT
  * #define USE_GTSAM_POSE_CONCEPT
  * #define USE_SIMPLE_POINT_CONCEPT