Clean up the code. Now we only use gtsam::Pose2
Since we already solved the compilation error,
we will not deal with other data-types for
the Point concept. This is less generic, but
now we want to keep the things simple and
easy to read. If tomorrow we see that is a great
tool, and that it would be nice to have it written
in a more generic way, we will go ahead and add
support for different data types(at least for gtsam)